annotate only the mutagenized ectodomain as the gene
We primarily want to analyze mutations in the ectodomain, which is the only
region mutagenized. There are some mutations in F upstream and downstream
of the ectodomain, but those mutations are sparse and often at low frequency.
We both are less interested in them, and they are mostly in multi-mutants
where global epistasis models might not fit things well.

Therefore, adjusted the PacBio amplicon and parse specs to only define the *gene*
to be the mutagenized ectodomain and to simply require the upstream and downstream
F regions not have mutations.

Also reduced the number of allowed mutations in the gene from 45 to a more
reasonable 15.

Then update the site numbering map so sequential numbering is sequential
in ectodomain.

Previously, the site numbering map just assigned everything to region "1",
instead assigned to F2, p27, or F1 in alignment w real biology of F.