add filters to remove high stdev measurements (#9)
This pull request addresses parts of #8 by adding filters to remove
measurements that are highly variable among replicates.

For the cell entry measurements, it adds an `effect_std: 1.6` and
`floor_for_effect_std: -2` setting. That means that we take the std
deviation of replicate measurements after putting a floor of -2,
and then remove any mutations with standard deviations bigger than that.

For escape, puts `escape_std: 1.25` to remove mutations with std deviation
in escape values that exceed that.

Also set a fixed min to the cell entry heatmap of -3, as it doesn't appear
to me that dynamic range below -3 is meaningful.