add `dms-viz` visualizations (#13)
This pull request adds `dms-viz` visualization of the cell
entry and antibody escape on prefusion and postfusion F,
as well as the co-crystal of prefusion F w nirsevimab.

It is set up so you can easily add additional DMS data and
structures via editing `data/dms_viz_config.yaml`.

The resulting JSON files are saved in the results, and are
linked in the HTML docs. But until repo is public, in order
to visualize you have to download those JSONs and then open
via *Upload* in `dms-viz`. When repo is public, we can
create links.